requirements.txt
transformers==4.36.2
tokenizers==0.15.0
huggingface-hub==0.20.3 
datasets==2.14.7
gradio==4.16.0
gradio_client==0.8.1
peft==0.4.0
bitsandbytes==0.41.0
httpx==0.24.0
einops==0.6.1
einops-exts==0.0.4
timm==0.6.13
pydantic==2.6.1
pillow==10.2.0
scikit-learn==1.2.2
xformers==0.0.22
easygui
open_clip_torch

#apex==0.9.10.dev0